% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corMatrix.R
\name{corMatrix}
\alias{corMatrix}
\title{Correlation matrix to be used for regression based imputatoin}
\usage{
corMatrix(dataframe, varpattern = "", debug = 0, test = 1, parallel = 0)
}
\arguments{
\item{varpattern}{A regular expression for subsetting variable names from input dataframe}

\item{debug}{debug mode; shows which models are running, the quality of predictions relative to original data, and any model errors. 1=progress notifications, errors and warnings.}

\item{test}{(deprecated)}

\item{parallel}{(deprecated)}
}
\value{
a list of two objects: $corMatrix == the correlation matrix between all variables given in the input dataframe, $df ==  a filtered version of the dataframe, that removes columns with no variance and applies any regex filters given.
}
\description{
Creates a dataframe with imputed values using either linear regression or lasso based models. For each variable in given data frame, the function finds the best correlated predictors (number of which is set by top_predictors), and uses these to construct models for predicting missing values.
}
\examples{
\dontrun{corMatrix(dataframe, varpattern="^c[mfhpktfvino]{1,2}[12345]", debug=1)}

}
