% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{imputation_ffc_init}
\alias{imputation_ffc_init}
\title{Data cleaning for logical and regression imputation}
\usage{
imputation_ffc_init(data = "", numeric = 1, dropna = 1, ageimpute = 1,
  meanimpute = 0)
}
\arguments{
\item{data}{location of data file}

\item{numeric}{convert all columns to numeric}

\item{dropna}{whether to convert missing values to NA}

\item{ageimpute}{perform logical imputation of mothers and fathers age across waves}

\item{meanimpute}{perform a mean imputation to help get better results from regression imputation; default off because regression imputation will perform this automatically.}
}
\value{
Dataframe with same dimensions as input, entirely numeric
}
\description{
Creates a dataframe with imputed values from
}
\examples{
\dontrun{imputation_ffc_init('background.csv', dropna=1, ageimpute=1, meanimpute=1)}

}
