% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_regression.R
\name{regImputation}
\alias{regImputation}
\title{Linear regression and lasso based imputation}
\usage{
regImputation(dataframe, matrix, method = "lm", parallel = 0,
  threshold = 0.4, top_predictors = 3, debug = 0, test = 0,
  degree = 1)
}
\arguments{
\item{method}{method for imputation (\code{"lm"} for ordinary least squares linear regression or \code{"lasso"} for lasso regularization)}

\item{parallel}{whether to use parallel processes (for MacOSX only at the moment)}

\item{threshold}{for selection of predictors based on correlation; values between 0 and 1.}

\item{top_predictors}{how many predictors to use in imputation prediction; more values can lead to better quality but more sparsely available predictions.}

\item{debug}{debug mode; shows which models are running, the quality of predictions relative to original data, and any model errors. 1=progress, errors and warnings, 2=progress,errors, warnings and prediction quality.}

\item{test}{test mode; runs on only the first 4 variables; helpful for trying out the function options before running full imputation.}

\item{degree}{the degree of polynomial effects to estimate (1=main effects only, 2=quadratic, 3=cubic, etc.)}
}
\value{
Dataframe containing imputed variables, with imputations performed only on missing values and retaining original data where available.
}
\description{
Creates a dataframe with imputed values using either linear regression or lasso based models. For each variable in given data frame, the function finds the best correlated predictors (number of which is set by top_predictors), and uses these to construct models for predicting missing values.
}
\examples{
\dontrun{regImputation(dataframe, method='polywog', varpattern="^c[mfhpktfvino]{1,2}[12345]", parallel=1, debug=1, test=1)}

}
