% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_regression.R
\name{regression_imputation}
\alias{regression_imputation}
\title{Linear regression and lasso based imputation}
\usage{
regression_imputation(dataframe, method = "lm", parallel = 0,
  threshold = 0.4, top_predictors = 3, constructed = 1, debug = 0,
  test = 0)
}
\arguments{
\item{method}{method for imputation (\code{"lm"} or \code{"polywog"})}

\item{parallel}{whether to use parallel processes}

\item{threshold}{for selection of predictors based on correlation; values between 0 and 1.}

\item{top_predictors}{how many predictors to use in imputation prediction; more values can lead to better quality but more sparsely available predictions.}

\item{constructed}{whether to return only constructed variables, or any variables passed to function; warning: slow if running on full dataset.}

\item{debug}{debug mode; shows which models are running, the quality of predictions relative to original data, and any model errors. 1=errors and warnings, 2=errors, warnings and prediction quality.}

\item{test}{test mode; runs on only the first 4 variables; helpful for trying out the function options before running full imputation.}
}
\value{
Dataframe containing imputed variables, with imputations performed only on missing values and retaining original data where available.
}
\description{
Creates a dataframe with imputed values using either linear regression or lasso based models. For each variable in given data frame, the function finds the best correlated predictors (number of which is set by top_predictors), and uses these to construct models for predicting missing values.
}
\examples{
\dontrun{regression_imputation(dataframe, method='polywog', parallel=1, debug=1, test=1)}

}
